refactor: extract MCP tool logic into mcp_tools/ package
- Extract 30 tool implementations from 2,067-line god function into 8 focused modules
- mcp_server.py: 2,543 → 1,543 lines (39% reduction)
- New mcp_tools/ package: scanning, analysis, compliance, sbom, registry, runtime, cloud, specialized
- @mcp.tool decorators stay in mcp_server.py (thin wrappers delegating to *_impl functions)
- Backward compat preserved: from agent_bom.mcp_server import create_mcp_server
- All 73 MCP tests pass, zero modifications to test files
