+++ src/agent_bom/mcp_tools/__init__.py
@@ -1,0 +1,6 @@
+"""MCP tool implementations — extracted from mcp_server.py for testability.
+
+Each sub-module contains the business logic for a group of MCP tools.
+The @mcp.tool registration remains in mcp_server.py; these modules
+provide the implementation functions that the thin wrappers call.
+"""
